fix(agent): preserve exact compaction request boundaries
