+++ tests/providers/test_openai_responses.py
@@ -743,11 +743,13 @@ class TestParseResponseOutput:
 
         assert compacted.provider_compaction_applied is True
         assert compacted.provider_compaction_state is not None
+        assert compacted.provider_compaction_scope == "current_request"
         assert responses_state_items(compacted.provider_compaction_state) == [
             {"type": "compaction", "encrypted_content": "opaque"},
         ]
         assert replayed.provider_compaction_applied is False
         assert replayed.provider_compaction_state is None
+        assert replayed.provider_compaction_scope is None
 
 
 class TestResponsesConversationState:
